refactor: server-only/client-only 적용으로 서버-클라이언트 경계 강화 (#210)
- CSR/SSR 혼동 방지 및 실행 위치 명시로 런타임 오류 예방
