+++ package.json
@@ -25,6 +25,7 @@
     "@reduxjs/toolkit": "^2.5.0",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.0",
+    "client-only": "^0.0.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.456.0",
     "next": "^15.2.4",
