+++ src/shared/api/index.ts
@@ -3,6 +3,7 @@ export { checkIpAllowed } from './check-ip/check-ip-access'
 export {
   type CheckIpRequest,
   type CheckIpResponse,
+  type ClientIpResult,
   CheckIpResponseSchema,
   IpSchema,
 } from './check-ip/schema'
